blog/requirements.txt
asgiref==3.6.0
Django==3.2.16
django-taggit==1.5.1
pytz==2022.7.1
sqlparse==0.4.3
